ensure "debug" is passed first
also ensure process.execArgv is forwarded

+++ index.js
@@ -1,49 +1,55 @@
 
 module.exports = false;
 
-if (!process.execArgv.length) {
+var args = [process.argv[1]];
+var argv = process.argv.slice(2);
+var debug = false;
+
+argv.forEach(function(arg) {
+  var flag = arg.split('=')[0];
+
+  switch (flag) {
+    case 'debug':
+      debug = true;
+      break;
+    case '--debug':
+    case '--debug-brk':
+    case '--expose-gc':
+    case '--gc-global':
+    case '--harmony':
+    case '--harmony-proxies':
+    case '--harmony-collections':
+    case '--harmony-generators':
+    case '--prof':
+      args.unshift(arg);
+      break;
+    default:
+      if (0 == arg.indexOf('--trace')) args.unshift(arg);
+      else args.push(arg);
+      break;
+  }
+});
+
+// has special args
+if (debug || args[0] != process.argv[1]) {
   var spawn = require('child_process').spawn;
 
-  var args = [process.argv[1]];
-  var argv = process.argv.slice(2);
+  // keep existing v8 args
+  args = process.execArgv.concat(args);
 
-  argv.forEach(function(arg) {
-    var flag = arg.split('=')[0];
+  // debug HAS to go first
+  if (debug) args.unshift('debug');
 
-    switch (flag) {
-      case 'debug':
-      case '--debug':
-      case '--debug-brk':
-      case '--expose-gc':
-      case '--gc-global':
-      case '--harmony':
-      case '--harmony-proxies':
-      case '--harmony-collections':
-      case '--harmony-generators':
-      case '--prof':
-        args.unshift(arg);
-        break;
-      default:
-        if (0 == arg.indexOf('--trace')) args.unshift(arg);
-        else args.push(arg);
-        break;
-    }
+  // recreate the process with corrected argument ordering
+  var proc = spawn(process.argv[0], args, { customFds: [ 0, 1, 2 ] });
+
+  proc.on('exit', function(code, signal) {
+    // ensure the main process exits with the same code
+    process.on('exit', function(){
+      if (signal) process.kill(process.pid, signal);
+      else process.exit(code);
+    });
   });
 
-  // has special args
-  if (args[0] != process.argv[1]) {
-    var proc = spawn(process.argv[0], args, { customFds: [ 0, 1, 2 ] });
-
-    proc.on('exit', function(code, signal) {
-      process.on('exit', function(){
-        if (signal) {
-          process.kill(process.pid, signal);
-        } else {
-          process.exit(code);
-        }
-      });
-    });
-
-    module.exports = true;
-  }
+  module.exports = true;
 }
